tests are ok... need edge cases

+++ packages/server/crud.js
@@ -62,18 +62,46 @@ const read = db => async (req, res) => {
 
 const update = (db) => async (req, res) => {
     const doc = req.body;
-    let replaceOne = await db.replaceOne({_id: new ObjectId(req.params.id)}, doc).catch(err => console.log(err));
-    res.json(doc);
+    const result = await db.findOne({_id: new ObjectId(req.params.id)})
+    if (result !== null) {
+        if(req.headers.authorization === undefined || result._authorized_readers.count === 0 ||result._authorized_readers.includes(getSub(req.headers.authorization))){
+            await db.replaceOne({_id: new ObjectId(req.params.id)}, doc).catch(err => console.log(err));
+            res.json(doc);
+        } else {
+            res.status(403);
+            res.json({})
+        }
+    } else {
+        res.status(404);
+        res.json({})
+    }
 };
 
 const remove = (db) => async (req, res) => {
-    await db.deleteOne({_id: new ObjectId(req.params.id)}).catch(err => console.log(err))
-
-    res.json({"deleted": req.params.id})
+    const result = await db.findOne({_id: new ObjectId(req.params.id)})
+    if (result !== null) {
+        if(req.headers.authorization === undefined || result._authorized_readers.count === 0 ||result._authorized_readers.includes(getSub(req.headers.authorization))){
+            await db.deleteOne({_id: new ObjectId(req.params.id)}).catch(err => console.log(err))
+            res.json({"deleted": req.params.id})
+        } else {
+            res.status(403);
+            res.json({})
+        }
+    } else {
+        res.status(404);
+        res.json({})
+    }
 };
 
 const list = (db) => async (req, res) => {
-    const results = await db.find().toArray();
+    let results;
+
+    if(req.headers.authorization === undefined){
+        results = await db.find().toArray();
+    }else {
+        let sub = getSub(req.headers.authorization);
+        results = await db.find({_authorized_readers: sub}).toArray();
+    }
     res.json(results);
 }
